+++ packages/node/src/api.ts
@@ -3,9 +3,13 @@ import axios, { AxiosInstance } from "axios";
 import { Workspace } from "./workspace";
 import { Releases } from "./releases";
 import { Subscribers } from "./subscribers";
-import { Roadmap } from "./roadmap";
+import { Milestones } from "./milestones";
 import { Webhooks } from "./webhooks";
 import { IncomingWebhooks } from "./incoming-webhooks";
+import { Projects } from "./projects";
+import { Labels } from "./labels";
+import { Contributors } from "./contributors";
+import { SubscriberLists } from "./subscriber-lists";
 
 export default class OnsetAPI {
   protected version = "v1";
@@ -33,12 +37,24 @@ export default class OnsetAPI {
     return new Releases(this.client);
   }
 
+  get milestones() {
+    return new Milestones(this.client);
+  }
+
+  /**
+   * @deprecated Use `milestones` instead. The `/roadmap` endpoint has been
+   * renamed to `/milestones` to match the Onset API reference.
+   */
   get roadmap() {
-    return new Roadmap(this.client);
+    return new Milestones(this.client);
   }
 
   get subscribers() {
     return new Subscribers(this.client);
+  }
+
+  get subscriberLists() {
+    return new SubscriberLists(this.client);
   }
 
   get webhooks() {
@@ -47,5 +63,17 @@ export default class OnsetAPI {
 
   get incomingWebhooks() {
     return new IncomingWebhooks(this.client);
+  }
+
+  get projects() {
+    return new Projects(this.client);
+  }
+
+  get labels() {
+    return new Labels(this.client);
+  }
+
+  get contributors() {
+    return new Contributors(this.client);
   }
 }